embed-editor page formattingtools section added and a new bg-overlay theme created.
