--- src/app/editor/page.tsx
@@ -1,10 +1,0 @@
-import EmbedEditorPageClient from "./EmbedEditorPageClient"
-
-export const metadata = {
-  title: "Smart Editor | DocStar",
-  description: "A powerful embedded editor with real-time collaboration and formatting tools."
-}
-
-export default function Page() {
-  return <EmbedEditorPageClient />
-}
