+++ src/scss/_utility.scss
@@ -13,6 +13,10 @@
   background-color: var(--app-bg-secondary);
 }
 
+.theme-bg-overlay {
+  background-color: var(--app-bg-overlay);
+}
+
 .theme-bg {
   background-color: var(--app-text);
 }
